requirements.txt
absl-py                 1.4.0
addict                  2.4.0
aiohttp                 3.8.4
aiosignal               1.3.1
argcomplete             3.0.5
asttokens               2.2.1
async-timeout           4.0.2
attrs                   22.2.0
backcall                0.2.0
beautifulsoup4          4.12.2
boto                    2.49.0
cachetools              5.3.0
certifi                 2022.12.7
cffi                    1.15.1
chamfer                 2.0.0
charset-normalizer      3.1.0
click                   8.1.3
colorama                0.4.6
comm                    0.1.3
ConfigArgParse          1.5.3
contourpy               1.0.7
cosine-annealing-warmup 2.0
crcmod                  1.7
cryptography            40.0.1
cycler                  0.11.0
dash                    2.9.2
dash-core-components    2.0.0
dash-html-components    2.0.0
dash-table              5.0.0
debugpy                 1.6.7
decorator               5.1.1
docker-pycreds          0.4.0
easydict                1.9
entrypoints             0.3
executing               1.2.0
fasteners               0.18
fastjsonschema          2.16.3
filelock                3.11.0
flake8                  3.7.9
Flask                   2.2.3
fonttools               4.39.3
frozenlist              1.3.3
ftfy                    6.0.1
future                  0.18.3
gcs-oauth2-boto-plugin  3.0
gdown                   4.7.1
gitdb                   4.0.10
GitPython               3.1.31
google-apitools         0.5.32
google-auth             2.17.2
google-auth-oauthlib    1.0.0
google-reauth           0.1.1
grpcio                  1.53.0
gsutil                  5.21
h5py                    3.6.0
httplib2                0.20.4
idna                    3.4
importlib-metadata      6.3.0
importlib-resources     5.12.0
ipykernel               6.22.0
ipython                 8.12.0
ipywidgets              8.0.6
isort                   4.3.21
itsdangerous            2.1.2
jedi                    0.18.2
Jinja2                  3.1.2
joblib                  1.2.0
jsonschema              4.17.3
jupyter_client          8.1.0
jupyter_core            5.3.0
jupyterlab-widgets      3.0.7
kiwisolver              1.4.4
littleutils             0.2.2
lmdb                    1.3.0
Markdown                3.4.3
MarkupSafe              2.1.2
matplotlib              3.7.1
matplotlib-inline       0.1.6
mccabe                  0.6.1
monotonic               1.6
multidict               6.0.4
multimethod             1.9.1
nbformat                5.5.0
nest-asyncio            1.5.6
numpy                   1.24.2
oauth2client            4.1.3
oauthlib                3.2.2
ogb                     1.3.6
open3d                  0.16.0
opencv-python           4.8.0.76
outdated                0.2.2
packaging               23.0
pandas                  2.0.0
parso                   0.8.3
pathtools               0.1.2
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.5.0
pip                     23.0.1
pkgutil_resolve_name    1.3.10
platformdirs            3.2.0
plotly                  5.14.1
pointnet2-cuda          0.0.0
promise                 2.3
prompt-toolkit          3.0.38
protobuf                3.20.3
psutil                  5.9.4
ptyprocess              0.7.0
pure-eval               0.2.2
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycodestyle             2.5.0
pycparser               2.21
pyflakes                2.1.1
Pygments                2.15.0
pyOpenSSL               23.1.1
pyparsing               3.0.9
pyquaternion            0.9.9
pyrsistent              0.19.3
PySocks                 1.7.1
pyte                    0.8.1
python-dateutil         2.8.2
python-dotenv           1.0.0
pytz                    2023.3
pyu2f                   0.1.5
PyYAML                  6.0
pyyaml_env_tag          0.1
pyzmq                   25.0.2
regex                   2022.10.31
requests                2.28.2
requests-oauthlib       1.3.1
retry-decorator         1.1.1
rsa                     4.7.2
scikit-learn            1.2.2
scipy                   1.10.1
sentry-sdk              1.19.1
setproctitle            1.3.2
setuptools              65.6.3
shortuuid               1.0.11
six                     1.16.0
smmap                   5.0.0
soupsieve               2.4
stack-data              0.6.2
tabulate                0.9.0
tb-nightly              2.13.0a20230410
tenacity                8.2.2
tensorboard-data-server 0.7.0
tensorboard-plugin-wit  1.8.1
termcolor               1.1.0
threadpoolctl           3.1.0
timm                    0.4.12
tornado                 6.2
tqdm                    4.65.0
traitlets               5.9.0
typing_extensions       4.5.0
tzdata                  2023.3
urllib3                 1.26.15
wandb                   0.13.3
wcwidth                 0.2.6
Werkzeug                2.2.3
wheel                   0.38.4
widgetsnbextension      4.0.7
wilds                   1.2.2
yacs                    0.1.8
yapf                    0.29.0
yarl                    1.8.2
zipp                    3.15.0